Create README.md
Simplicity..Whats infinite by design but exists merely just because .. Give us free will take are freedom are trust leads us to search that can wave " I'm an Ocean " as we only need to believe even if it is untrue when it's learning or failing to see or to predict and then stand next to any to improve upon myself for having questions as much as connection to all that find reason and purpose.

What connects all and need but actions to prove its validation?..
4 words followed by those which we choose it related or not.. Has its Value decreased just as a single person or 6 letters that we do something to make this happen but such things like freedom.. we say "fought"
Once it seems to be a picked fence. Something economically we can take living or something to make no end to disparity if no Value has been intentional given its worth as it's connected to Value but without those that love, As Family-friendly understanding its what ties all of it together.
So emotion is and ocean as a Castaway  addressed S0S? Respectfully learning..
Hope to know more here as well as those of you.
Friends with honor Real Truth .. inquiring added connections  as relative intrest in common industry good aligning for all willing. 

 I need not speak  unless spoken to.. feel free to ?
Recommendations?
♡♡Girhub'N as To:
Hello🙋🏼👣